ref(Global): move `IGameDataProvider` to a separate file
