+++ Assets/Gothic-UnZENity-Core/Scripts/GameGlobals.cs
@@ -7,21 +7,6 @@ using GUZ.Core.World;
 
 namespace GUZ.Core
 {
-	public interface IGlobalDataProvider
-	{
-		public GameConfiguration Config { get; }
-		public GameSettings Settings { get; }
-		public SkyManager Sky { get; }
-		public GameTime Time { get; }
-		public RoutineManager Routines { get; }
-		public TextureManager Textures { get; }
-		public GUZSceneManager Scene { get; }
-		public FontManager Font { get; }
-		public StationaryLightsManager Lights { get; }
-		public VobMeshCullingManager MeshCulling { get; }
-		public VobSoundCullingManager SoundCulling { get; }
-	}
-
 	public static class GameGlobals
 	{
 		public static IGlobalDataProvider Instance;
