+++ Assets/Gothic-UnZENity-Core/Scripts/IGlobalDataProvider.cs
@@ -1,0 +1,22 @@
+using GUZ.Core.Manager;
+using GUZ.Core.Manager.Culling;
+using GUZ.Core.Manager.Settings;
+using GUZ.Core.World;
+
+namespace GUZ.Core
+{
+	public interface IGlobalDataProvider
+	{
+		public GameConfiguration Config { get; }
+		public GameSettings Settings { get; }
+		public SkyManager Sky { get; }
+		public GameTime Time { get; }
+		public RoutineManager Routines { get; }
+		public TextureManager Textures { get; }
+		public GUZSceneManager Scene { get; }
+		public FontManager Font { get; }
+		public StationaryLightsManager Lights { get; }
+		public VobMeshCullingManager MeshCulling { get; }
+		public VobSoundCullingManager SoundCulling { get; }
+	}
+}
