+++ Assets/Gothic-UnZENity-Core/Scripts/IGlobalDataProvider.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: d535c293054d42abba54956de5cdc049
+timeCreated: 1719335284
